New tests to cover CG, DG and mixed element spaces
